Fixed crc checksum issue in US_DataIO for maximum portability

+++ trunk/utils/us_crc.cpp
@@ -1,7 +1,7 @@
 #include "us_crc.h"
 
-unsigned long US_Crc::crc32( 
-      unsigned long crc, const unsigned char* buf, unsigned int len )
+quint32 US_Crc::crc32( 
+        quint32 crc, const unsigned char* buf, unsigned int len )
 {
    if ( buf == 0 ) return 0UL;
